% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_shannon.R
\name{my_shannon}
\alias{my_shannon}
\title{Function Shannon}
\usage{
my_shannon(x)
}
\arguments{
\item{x}{Avector of numeric values representig the
abundance species in a community}
}
\value{
Shannon diversity value
}
\description{
Function Shannon
}
\examples{
com <- c(1, 2, 3)
my_shannon(com)
}
